pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "labelrag"
version = "0.0.3"
description = "A label-driven RAG pipeline built on top of paralabelgen."
readme = "README.md"
license = { text = "MIT" }
authors = [
  { name = "huruilizhen", email = "huruilizhen@gmail.com" },
]
requires-python = ">=3.11"
dependencies = [
  "numpy>=1.26.0",
  "paralabelgen==0.2.2",
  "sentence-transformers>=3.0.0",
]
keywords = ["rag", "retrieval", "labels", "evaluation", "paragraphs", "llm"]
classifiers = [
  "Development Status :: 2 - Pre-Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "Topic :: Software Development :: Libraries :: Python Modules",
]

[project.urls]
Homepage = "https://github.com/HuRuilizhen/labelrag"
Repository = "https://github.com/HuRuilizhen/labelrag"
Issues = "https://github.com/HuRuilizhen/labelrag/issues"

[project.optional-dependencies]
dev = [
  "build>=1.2.0",
  "pyright>=1.1.390",
  "pytest>=8.3.0",
  "pytest-cov>=6.0.0",
  "ruff>=0.11.0",
  "twine>=6.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/labelrag"]

[tool.hatch.build.targets.sdist]
only-include = [
  "/src/labelrag",
  "/README.md",
  "/LICENSE",
  "/pyproject.toml",
]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
addopts = [
  "--strict-config",
  "--strict-markers",
]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pyright]
include = ["src", "tests", "examples"]
pythonVersion = "3.11"
typeCheckingMode = "strict"
# `paralabelgen` does not currently ship type stubs, so this remains an
# intentional deviation from the sibling `labelgen` repository.
reportMissingTypeStubs = false
venvPath = "."
venv = ".venv"
